+++ src/diffusers/models/unet_2d_condition.py
@@ -191,6 +191,6 @@ class UNet2DConditionModel(ModelMixin, ConfigMixin, MPSWarmupMixin):
     def warmup_inputs(self, batch_size) -> Tuple:
         batch_size = 1 if batch_size is None else batch_size
         w_sample = torch.randn((batch_size, self.in_channels, 64, 64))
-        t = torch.tensor([10], dtype=torch.long)
+        t = torch.tensor([10], dtype=torch.int32)
         w_encoded = torch.rand((batch_size, 77, 768))
         return (w_sample, t, w_encoded)
